requirements.txt
sarvamai
voice-cloning
librosa
soundfile
python-dotenv
elevenlabs
apscheduler